requirements.txt
youtube-transcript-api
openai
google-generativeai
python-dotenv
rich
pyperclip
